Move functions and stringcache to eager
