+++ py-polars/polars/__init__.py
@@ -4,10 +4,9 @@ from polars.eager.series import Series, wrap_s  # noqa: F401
 from polars.lazy.expr import Expr, wrap_expr  # noqa: F401
 from polars.lazy.frame import LazyFrame, wrap_ldf  # noqa: F401
 
-from . import datatypes, eager, functions, io, lazy
+from . import datatypes, eager, io, lazy
 from .datatypes import *  # noqa: F401, F403
 from .eager import *  # noqa: F401, F403
-from .functions import *  # noqa: F401, F403
 from .io import *  # noqa: F401, F403
 from .lazy import *  # noqa: F401, F403
 
@@ -19,6 +18,4 @@ try:
 except ImportError:
     pass
 
-__all__ = (
-    datatypes.__all__ + eager.__all__ + functions.__all__ + io.__all__ + lazy.__all__
-)
+__all__ = datatypes.__all__ + eager.__all__ + io.__all__ + lazy.__all__
