+++ py-polars/polars/eager/__init__.py
@@ -1,5 +1,7 @@
-from . import frame, series
+from . import frame, functions, series, string_cache
 from .frame import *  # noqa: F401, F403
+from .functions import *  # noqa: F401, F403
 from .series import *  # noqa: F401, F403
+from .string_cache import *  # noqa: F401, F403
 
-__all__ = frame.__all__ + series.__all__
+__all__ = frame.__all__ + functions.__all__ + series.__all__ + string_cache.__all__
